backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
psycopg[binary]>=3.1
alembic>=1.13
pydantic-settings>=2.2
